+++ lib/hephaestus_ecto/migrations/postgres.ex
@@ -14,7 +14,7 @@ defmodule HephaestusEcto.Migrations.Postgres do
   use Ecto.Migration
 
   @initial_version 1
-  @current_version 2
+  @current_version 3
   @default_prefix "public"
 
   @valid_prefix ~r/^[a-z_][a-z0-9_]*$/