+++ priv/repo/migrations/20260412000000_upgrade_workflow_instances_to_v03.exs
@@ -1,0 +1,6 @@
+defmodule HephaestusEcto.TestRepo.Migrations.UpgradeWorkflowInstancesToV03 do
+  use Ecto.Migration
+
+  def up, do: HephaestusEcto.Migration.up(version: 3)
+  def down, do: HephaestusEcto.Migration.down(version: 2)
+end